Modify the type error definition

+++ src/unit_byte.hpp
@@ -332,7 +332,7 @@ private:
     uint8_t _addr;
     uint8_t _scl;
     uint8_t _sda;
-    uint8_t _speed;
+    uint32_t _speed;
 
     /**
      * @brief Writes multiple bytes to a specified register.
